[AUDIO] Provide a generic Android macro to replace patterns in files
There is a need for a generic Android macro to replace a list of
patterns in files (typicall a key/value list). Currently, one macro
is defined for each single pattern.

This patch provides a new macro called BUILD_REPLACE_PATTERNS_IN_FILE
that takes in parameter a list of pairs that will be used to replace
patterns by their associated value in the input file.
The patterns list shall be named REPLACE_PATTERNS and be composed of
space separated pairs, following this syntax:
        REPLACE_PATTERNS := pattern_1=value1 pattern2=value2

Change-Id: Ic188fed06f41adf685a814e3d981aa6b4b1b28ee
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-10924
Signed-off-by: Frederic Boisnard <frederic.boisnard@intel.com>
Reviewed-on: https://android.intel.com/343663
Reviewed-by: Ochoa, Alejandro <alejandro.ochoa@intel.com>
Tested-by: Ochoa, Alejandro <alejandro.ochoa@intel.com>